requirements.txt
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.1.1
google-genai==1.39.1
ja_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ja_core_news_sm-3.8.0/ja_core_news_sm-3.8.0-py3-none-any.whl#sha256=65693016b7cf33ffa845af3523a9c9b8d9aabae85dfc2e46731513a5e03def04
python-dotenv==1.1.1
spacy==3.8.11
spacy-stanza==1.0.4
stanza==1.6.1
yt-dlp==2025.10.22

faster-whisper==1.2.1
ctranslate2==4.6.1
huggingface-hub==1.1.5
onnxruntime==1.23.2
numpy==1.26.4
tokenizers==0.22.1
packaging==25.0
protobuf==6.33.1
tqdm==4.67.1
typing_extensions==4.15.0

torch==2.6.0
torchaudio==2.6.0
soundfile==0.13.1
nvidia-cudnn-cu12==9.1.0.70
nvidia-cublas-cu12==12.1.3.1
